Rewrite thread, debugger and state system for gb
I'm still not fully happy with how the pause system relies on the Debugger class but I'm okay with the "hack" for now

+++ src/core/PPU.h
@@ -120,7 +120,7 @@ public:
 	static inline bool windowEnabled = true;
 	static inline bool spritesEnabled = true;
 
-	size_t framesPresented;
+	u16 framesPresented;
 
 	std::mutex vblank_m;
